chore: move nyt games to own dir
